🗃️ Seperate DB module from nestjs
