+++ apps/api/src/core/auth/auth.controller.ts
@@ -29,7 +29,7 @@ export class AuthController {
 
 
         });
-        return this.authService.login(req.user);
+        // return this.authService.login(req.user);
     }
 
 }
