+++ apps/api/src/core/auth/dto/requestUserDto.ts
@@ -1,6 +1,6 @@
-import { User } from '@/drizzle/schema';
+import { NewUser } from '@template/shared';
 import { Request } from 'express';
 
 export interface RequestUserDto extends Request {
-    user: User;
+    user: NewUser;
 }
