--- apps/api/src/drizzle/types/drizzle.d.ts
@@ -1,4 +1,0 @@
-import * as schema from '../schema';
-import { NodePgDatabase } from 'drizzle-orm/node-postgres';
-
-export type DrizzleDB = NodePgDatabase<typeof schema>;
