+++ packages/shared/drizzle.config.ts
@@ -3,7 +3,7 @@ import 'dotenv/config';
 import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
-    schema: './src/drizzle/schema.ts',
+    schema: './src/db/schema.ts',
     dialect: 'postgresql',
     dbCredentials: {
         url: process.env.DATABASE_URL!,
