+++ packages/shared/src/main.ts
@@ -1,0 +1,20 @@
+// contract.ts
+
+import { initContract } from '@ts-rest/core';
+import { postContract } from './routers/posts';
+import { userContract } from './routers/users';
+
+const c = initContract();
+
+export const contract = c.router(
+    {
+        posts: postContract,
+        users: userContract
+    },
+    { pathPrefix: '/api', strictStatusCodes: true }
+);
+
+
+// export * from './db';
+export * from './db/schema';
+export * from './db/types';
